+++ apps/app-backend/src/config.ts
@@ -68,7 +68,11 @@ export const PaginationConfig = t.type({
   ),
   PAGINATION_MAX_OFFSET: withDefault(
     IntegerFromString.pipe(NonNegativeInteger),
-    "500" as unknown as NonNegativeInteger
+    "100" as unknown as NonNegativeInteger
+  ),
+  PAGINATION_MAX_OFFSET_AI_SEARCH: withDefault(
+    IntegerFromString.pipe(NonNegativeInteger),
+    "100000" as unknown as NonNegativeInteger
   ),
 });